+++ example/lib/example5_themes_playground/theme/flex_theme_light.dart
@@ -312,7 +312,7 @@ FlexColorScheme flexColorSchemeLight(ThemeController controller, Color source) {
             drawerBackgroundSchemeColor: controller.drawerBackgroundSchemeColor,
             drawerWidth: controller.drawerWidth,
             drawerIndicatorWidth: controller.drawerIndicatorWidth,
-            drawerIndicatorBorderRadius: controller.drawerIndicatorBorderRadius,
+            drawerIndicatorRadius: controller.drawerIndicatorBorderRadius,
             drawerIndicatorSchemeColor: controller.drawerIndicatorSchemeColor,
             drawerIndicatorOpacity: controller.drawerIndicatorOpacity,
             // BottomSheet settings.